Trades.Leader - given info source
Inspired by frathons hedgehog.
Trying a different approach to the problem.

Leader gets all the information and tries to analyze the data
the conclusions is then served for the Trades.Junior to use in the
trading.

There is a Leader for every instrument.
There is many Juniors for every leader, depending on market.

This commit ripps up some previous work.

When app is started, a Leader will start a trading stream and logging
all trade events to stdout.